Re-implement passwdok using libssh2:
the supplied password will be used to attempt to authenticate an ssh
ssh session with a server listening on the localhost.

This method has the advantage of not relying on the binary being
installed setuid and lets the ssh server handle anything that
requires special permissions in a secure way.

Also: use functions from skalibs/strerr2.h instead of stdio for error
and warning messages (mostly since I prefer the skalibs interface).